webserver/requirements.txt
# requirements.txt

# TODO sort this out later... 
# It would probably be in our best interest to have version numbers
Flask
gunicorn
boto3
psycopg2-binary
datetime
python-dotenv
Flask-JWT-Extended
# TODO add these for sqlite, however, I can't get them to work on my machine
sqlite-utils
# json
ultralytics
matplotlib
ultralytics-thop
opencv-python-headless
onnxruntime
pybind11
langchain_openai
langchain_core
langchain_community
langchain
openai
sentence_transformers
xformers
openai-whisper